requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
aiohttp-socks==0.3.4
alabaster==0.7.12
alembic==1.4.0
altair==4.0.1
aniso8601==8.0.0
appdirs==1.4.4
applaunchservices==0.2.1
appnope==0.1.0
appscript==1.0.1
argh==0.26.2
arviz==0.8.3
asn1crypto==1.3.0
astor==0.8.1
astroid==2.3.3
astropy==4.0.1.post1
astroquery==0.4
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
base58==2.0.0
beautifulsoup4==4.9.0
bitarray==1.2.1
bkcharts==0.2
bleach==3.1.4
blinker==1.4
bokeh==2.0.2
boto==2.49.0
boto3==1.12.39
botocore==1.15.39
Bottleneck==1.3.2
cachetools==3.1.1
cchardet==2.1.5
Cerberus==1.3.2
certifi==2020.4.5.1
cffi==1.14.0
cftime==1.1.3
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
cloudpickle==1.4.1
colorama==0.4.3
configparser==5.0.0
contextlib2==0.6.0.post1
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==2.9.2
cycler==0.10.0
cymem==1.31.2
Cython==0.29.17
cytoolz==0.9.0.1
dask==2.10.1
decorator==4.4.2
defusedxml==0.6.0
descartes==1.1.0
diff-match-patch==20181111
dill==0.2.9
distlib==0.3.0
distributed==2.10.0
docutils==0.15.2
elasticsearch==7.5.1
emoji==0.5.4
entrypoints==0.3
enum-compat==0.0.3
et-xmlfile==1.0.1
Eve==1.1
Events==0.3
fake-useragent==0.1.11
fastcache==1.1.0
fastprogress==0.2.3
filelock==3.0.12
Fiona==1.8.13.post1
flake8==3.8.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-HTTPAuth==4.0.0
Flask-Migrate==2.5.2
Flask-RESTful==0.3.8
flask-restful-swagger==0.20.1
Flask-Sphinx-Themes==1.0.2
Flask-SQLAlchemy==2.4.1
Flask-Toastr==0.5.4
fsspec==0.7.1
future==0.18.2
gensim==3.8.0
geographiclib==1.50
geopandas==0.7.0
geopy==1.21.0
gevent==1.4.0
glob2==0.7
gmpy2==2.0.8
google-api-core==1.17.0
google-auth==1.14.1
google-cloud-core==1.3.0
google-cloud-storage==1.28.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
googletransx==2.4.2
greenlet==0.4.15
gunicorn==20.0.4
h3==3.6.3
h5py==2.10.0
hatebase==0.6
HeapDict==1.0.1
html5lib==1.0.1
hypothesis==5.11.0
idna==2.9
imageio==2.8.0
imagesize==1.2.0
imbalanced-learn==0.6.1
importlib-metadata==1.6.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4.1
jedi==0.15.2
Jinja2==2.11.2
jmespath==0.9.4
joblib==0.15.1
json-tricks==3.14.0
json5==0.9.4
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==1.1.4
jupyterlab-server==1.1.4
keyring==21.1.1
kiwisolver==1.2.0
langcodes==1.4.1
langdetect==1.0.8
lazy-object-proxy==1.4.3
libarchive-c==2.9
llvmlite==0.32.1
locket==0.2.0
lxml==4.5.0
Mako==1.1.1
marisa-trie==0.7.5
MarkupSafe==1.1.1
marshmallow==3.6.0
marshmallow-jsonapi==0.23.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.2
more-itertools==8.3.0
mpldatacursor==0.7.1
mpmath==1.1.0
msgpack==0.5.6
msgpack-numpy==0.4.4.3
multidict==4.7.5
multipledispatch==0.6.0
munch==2.5.0
murmurhash==0.28.0
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.6
netCDF4==1.5.3
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.49.1
numexpr==2.7.1
numpy==1.18.1
numpydoc==0.9.2
olefile==0.46
openpyxl==3.0.3
packaging==20.3
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.2
partd==1.1.0
path==13.2.0
path.py==12.0.2
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
Pint==0.11
pkginfo==1.5.0.1
plac==0.9.6
plotly==4.6.0
pluggy==0.13.1
ply==3.11
preshed==1.0.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.4
psutil==5.7.0
psycopg2-binary==2.8.5
ptyprocess==0.6.0
publib==0.3.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pybbn==1.1.2
pycares==3.1.1
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycurl==7.43.0.5
pydeck==0.2.1
pydocstyle==4.0.1
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.4.2
pymc3==3.9.1
pymongo==3.10.1
pyodbc===4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==2.4.6
Pyphen==0.9.5
pyproj==2.4.2.post1
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.2
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.5.0
pytest-remotedata==0.3.2
python-crfsuite==0.9.7
python-dateutil==2.8.0
python-docx==0.8.10
python-editor==1.0.4
python-jsonrpc-server==0.3.4
python-language-server==0.31.9
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==18.1.1
QDarkStyle==2.8.1
QtAwesome==0.7.0
qtconsole==4.7.4
QtPy==1.9.0
radis==0.9.25
rdflib==4.2.2
redis==3.5.2
regex==2018.7.11
requests==2.23.0
retrying==1.3.3
rope==0.17.0
rsa==4.0
Rtree==0.9.4
ruamel-yaml==0.15.87
s3transfer==0.3.3
schedule==0.6.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.1
selenium==3.141.0
Send2Trash==1.5.0
shap==0.35.0
Shapely==1.7.0
simplegeneric==0.8.1
simplejson==3.17.0
singledispatch==3.4.0.3
six==1.15.0
sklearn-crfsuite==0.3.6
smart-open==2.0.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==2.0.1
spacy==2.0.12
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.1
spyder==4.1.3
spyder-kernels==1.9.1
SQLAlchemy==1.3.17
stanfordnlp==0.2.0
statsmodels==0.11.1
streamlit==0.55.0
sympy==1.5.1
tables==3.6.1
tabulate==0.8.7
tblib==1.6.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
textstat==0.6.0
Theano==1.0.4
thinc==6.10.3
toml==0.10.0
toolz==0.10.0
torch==1.4.0
torchvision==0.5.0
tornado==5.1.1
tqdm==4.46.0
traitlets==4.3.3
twint==2.1.15
typing-extensions==3.7.4.1
tzlocal==2.0.0
ujson==1.35
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.8
vaderSentiment==3.3.2
validators==0.14.2
vega-datasets==0.8.0
virtualenv==20.0.25
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordfreq==2.2.1
wrapt==1.10.11
wurlitzer==2.0.0
xarray==0.15.1
xgboost==1.1.0
xlrd==1.2.0
XlsxWriter==1.2.8
xlwings==0.19.4
xlwt==1.3.0
yapf==0.28.0
yarl==1.4.2
yellowbrick==1.1
zict==2.0.0
zipp==3.1.0
